+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Support compile-time automatic instrumentation of packages
+  that don't import `errtrace` by specifying the flag `unsafe-packages`.
+  This still requires at least one import of `errtrace` in the binary.
+
 ## 0.4.0 - 2025-07-21
 
 This release supports compile-time rewriting of source files via `toolexec`.